trending-agent/requirements.txt
google-adk
google-genai
a2a-sdk
uvicorn
python-dotenv
